Refactor Distributed Tests (#2180)
Refactor Distributed unit tests
