+++ tests/unit/ops/sparse_attention/test_sparse_attention.py
@@ -13,146 +13,6 @@ if not deepspeed.ops.__compatible_ops__[SparseAttnBuilder.NAME]:
                 allow_module_level=True)
 
 
-def test_sparse_attention_module_availability():
-    return True
-    try:
-        from deepspeed.ops import sparse_attention  # noqa: F401
-    except ImportError:
-        print("Sparse Attention Module is not installed!")
-        return False
-    return True
-
-
-def test_matmul_module_availability():
-    return True
-    try:
-        from deepspeed.ops.sparse_attention.matmul import MatMul  # noqa: F401
-    except ImportError:
-        print("Sparse MatMul Module is not installed!")
-        return False
-    return True
-
-
-def test_softmax_module_availability():
-    return True
-    try:
-        from deepspeed.ops.sparse_attention.softmax import Softmax  # noqa: F401
-    except ImportError:
-        print("Sparse Softmax Module is not installed!")
-        return False
-    return True
-
-
-def test_sparsityconfig_module_availability():
-    return True
-    try:
-        from deepspeed.ops.sparse_attention import SparsityConfig  # noqa: F401
-    except ImportError:
-        print("SparsityConfig Module is not installed!")
-        return False
-    return True
-
-
-def test_densesparsityconfig_module_availability():
-    return True
-    try:
-        from deepspeed.ops.sparse_attention import DenseSparsityConfig  # noqa: F401
-    except ImportError:
-        print("DenseSparsityConfig Module is not installed!")
-        return False
-    return True
-
-
-def test_fixedsparsityconfig_module_availability():
-    return True
-    try:
-        from deepspeed.ops.sparse_attention import FixedSparsityConfig  # noqa: F401
-    except ImportError:
-        print("FixedSparsityConfig Module is not installed!")
-        return False
-    return True
-
-
-def test_variablesparsityconfig_module_availability():
-    return True
-    try:
-        from deepspeed.ops.sparse_attention import VariableSparsityConfig  # noqa: F401
-    except ImportError:
-        print("VariableSparsityConfig Module is not installed!")
-        return False
-    return True
-
-
-def test_bigbirdsparsityconfig_module_availability():
-    return True
-    try:
-        from deepspeed.ops.sparse_attention import BigBirdSparsityConfig  # noqa: F401
-    except ImportError:
-        print("BigBirdSparsityConfig Module is not installed!")
-        return False
-    return True
-
-
-def test_bslongformersparsityconfig_module_availability():
-    return True
-    try:
-        from deepspeed.ops.sparse_attention import BSLongformerSparsityConfig  # noqa: F401
-    except ImportError:
-        print("BSLongformerSparsityConfig Module is not installed!")
-        return False
-    return True
-
-
-def test_localwindowsparsityconfig_module_availability():
-    return True
-    try:
-        from deepspeed.ops.sparse_attention import LocalSlidingWindowSparsityConfig  # noqa: F401
-    except ImportError:
-        print("LocalSlidingWindowSparsityConfig Module is not installed!")
-        return False
-    return True
-
-
-def test_sparseselfattention_module_availability():
-    return True
-    try:
-        from deepspeed.ops.sparse_attention import SparseSelfAttention  # noqa: F401
-    except ImportError:
-        print("SparseSelfAttention Module is not installed!")
-        return False
-    return True
-
-
-def test_bertsparseselfattention_module_availability():
-    return True
-    try:
-        from deepspeed.ops.sparse_attention import BertSparseSelfAttention  # noqa: F401
-    except ImportError:
-        print("BertSparseSelfAttention Module is not installed!")
-        return False
-    return True
-
-
-def test_sparseattentionutils_availability():
-    return True
-    try:
-        from deepspeed.ops.sparse_attention import SparseAttentionUtils  # noqa: F401
-    except ImportError:
-        print("SparseAttentionUtils Module is not installed!")
-        return False
-    return True
-
-
-def test_cpp_utils_availability():
-    return True
-    try:
-        from deepspeed.ops.sparse_attention import cpp_utils  # noqa: F401
-    except ImportError:
-        print("Sparse Attention cpp_utils Module is not installed!")
-        return False
-    return True
-
-
 def dense_to_sparse(w, mask, block):
     """Converts dense matrix with explicit zeros to sparse matrix
     """
