+++ tests/unit/runtime/test_autocast.py
@@ -3,15 +3,6 @@ import torch
 from deepspeed.runtime.zero.linear import LinearModuleForZeroStage3
 
 
-def _skip_autocast_test():
-    try:
-        from torch.cuda.amp import custom_fwd, custom_bwd  # noqa: F401
-    except (ImportError, AttributeError) as exp:
-        return True
-
-    return False
-
-
 @pytest.mark.parametrize('half_op', [False, True])
 def test_missing_amp_autocast(tmpdir, half_op):
     hidden_dim = 4
@@ -28,8 +19,7 @@ def test_missing_amp_autocast(tmpdir, half_op):
 
 @pytest.mark.parametrize('half_op', [False, True])
 def test_disable_autocast_linear(tmpdir, half_op):
-    if _skip_autocast_test():
-        pytest.skip("amp autocast is not available")
+    amp = pytest.importorskip("torch.cuda.amp")
 
     hidden_dim = 4
     if half_op:
@@ -39,7 +29,7 @@ def test_disable_autocast_linear(tmpdir, half_op):
         input = torch.randn(hidden_dim).cuda()
         ds_linear = LinearModuleForZeroStage3(hidden_dim, hidden_dim).cuda()
 
-    with torch.cuda.amp.autocast(False):
+    with amp.autocast(False):
         output = ds_linear(input)
         assert output.dtype == ds_linear.weight.dtype
 
@@ -54,8 +44,7 @@ def test_disable_autocast_linear(tmpdir, half_op):
                           (True,
                            True)])
 def test_autocast_linear(tmpdir, half_input, half_weight):
-    if _skip_autocast_test():
-        pytest.skip("amp autocast is not available")
+    amp = pytest.importorskip("torch.cuda.amp")
 
     hidden_dim = 4
     input = torch.randn(hidden_dim).cuda()
@@ -67,6 +56,6 @@ def test_autocast_linear(tmpdir, half_input, half_weight):
     if half_weight:
         ds_linear = ds_linear.half()
 
-    with torch.cuda.amp.autocast():
+    with amp.autocast():
         output = ds_linear(input)
         assert output.dtype == torch.half
