+++ tests/unit/runtime/test_data.py
@@ -2,8 +2,8 @@ from deepspeed.utils import RepeatingLoader
 import torch
 import pytest
 import deepspeed
-from .common import distributed_test
-from .simple_model import SimpleModel, args_from_dict, random_dataset
+from tests.unit.common import DistributedTest
+from tests.unit.simple_model import SimpleModel, random_dataset
 
 
 def test_repeating_loader():
@@ -25,26 +25,26 @@ def test_repeating_loader():
                            False),
                           (4,
                            False)])
-def test_dataloader_drop_last(tmpdir, train_batch_size, drop_last):
-    config_dict = {
-        "train_batch_size": train_batch_size,
-        "dataloader_drop_last": drop_last,
-        "steps_per_print": 1
-    }
-    args = args_from_dict(tmpdir, config_dict)
-    hidden_dim = 10
+class TestDataLoaderDropLast(DistributedTest):
+    world_size = 1
 
-    model = SimpleModel(hidden_dim)
+    def test(self, train_batch_size, drop_last):
+        config_dict = {
+            "train_batch_size": train_batch_size,
+            "dataloader_drop_last": drop_last,
+            "steps_per_print": 1
+        }
+        hidden_dim = 10
 
-    @distributed_test(world_size=[1])
-    def _test_dataloader_drop_last(args, model, hidden_dim):
+        model = SimpleModel(hidden_dim)
         optimizer = torch.optim.AdamW(params=model.parameters())
-        #TODO: Figure out why this breaks with cuda device
+        # TODO: no way to set DeepSpeedEngine.deepspeed_io params, need to use
+        # pin_memory=False for cuda device
         train_dataset = random_dataset(total_samples=50,
                                        hidden_dim=hidden_dim,
                                        device=torch.device('cpu'),
                                        dtype=torch.float32)
-        model, _, training_dataloader, _ = deepspeed.initialize(args=args,
+        model, _, training_dataloader, _ = deepspeed.initialize(config=config_dict,
                                                                 model=model,
                                                                 training_data=train_dataset,
                                                                 optimizer=optimizer)
@@ -54,5 +54,3 @@ def test_dataloader_drop_last(tmpdir, train_batch_size, drop_last):
             loss = model(x, y)
             model.backward(loss)
             model.step()
-
-    _test_dataloader_drop_last(args=args, model=model, hidden_dim=hidden_dim)
