+++ tests/unit/test_sparse_grads.py
@@ -1,42 +1,48 @@
 import torch
 import deepspeed
-from .common import distributed_test
+from tests.unit.common import DistributedTest
 
 import deepspeed.utils.groups as groups
 
 
-def test_sparse_adam(tmpdir):
-    config_dict = {"train_batch_size": 2, "steps_per_print": 1, "sparse_gradients": True}
+class Model(torch.nn.Module):
+    def __init__(self):
+        super().__init__()
+        self.emb = torch.nn.EmbeddingBag(10, 3, mode="sum", sparse=True)
+        self.linear = torch.nn.Linear(3, 1)
 
-    class Model(torch.nn.Module):
-        def __init__(self):
-            super().__init__()
-            self.emb = torch.nn.EmbeddingBag(10, 3, mode="sum", sparse=True)
-            self.linear = torch.nn.Linear(3, 1)
+    def forward(self, x, offsets):
+        return self.linear(self.emb(x, offsets))
 
-        def forward(self, x, offsets):
-            return self.linear(self.emb(x, offsets))
 
-    class Adam(torch.optim.Optimizer):
-        def __init__(self, dense_params, sparse_params):
-            super().__init__(dense_params + sparse_params, defaults={})
-            self.adam = torch.optim.Adam(dense_params)
-            self.adam_sparse = torch.optim.SparseAdam(sparse_params)
+class Adam(torch.optim.Optimizer):
+    def __init__(self, dense_params, sparse_params):
+        super().__init__(dense_params + sparse_params, defaults={})
+        self.adam = torch.optim.Adam(dense_params)
+        self.adam_sparse = torch.optim.SparseAdam(sparse_params)
 
-        @torch.no_grad()
-        def step(self, closure=None):
-            loss_1 = self.adam.step(closure)
-            loss_2 = self.adam_sparse.step(closure)
+    @torch.no_grad()
+    def step(self, closure=None):
+        loss_1 = self.adam.step(closure)
+        loss_2 = self.adam_sparse.step(closure)
 
-            if loss_1 is not None and loss_2 is not None:
-                return loss_1 + loss_2
-            return loss_1 or loss_2
+        if loss_1 is not None and loss_2 is not None:
+            return loss_1 + loss_2
+        return loss_1 or loss_2
 
-    model = Model()
-    optimizer = Adam(list(model.linear.parameters()), list(model.emb.parameters()))
 
-    @distributed_test(world_size=[2])
-    def _test(model, optimizer):
+class TestSparseAdam(DistributedTest):
+    world_size = 2
+
+    def test(self):
+        config_dict = {
+            "train_batch_size": 2,
+            "steps_per_print": 1,
+            "sparse_gradients": True
+        }
+
+        model = Model()
+        optimizer = Adam(list(model.linear.parameters()), list(model.emb.parameters()))
         engine, _, _, _ = deepspeed.initialize(model=model,
                                               optimizer=optimizer,
                                               config=config_dict)
@@ -64,5 +70,3 @@ def test_sparse_adam(tmpdir):
         ]
         for res in results:
             assert torch.allclose(res[0], res[1])
-
-    _test(model, optimizer)
