+++ tests/unit/utils/test_init_on_device.py
@@ -1,6 +1,6 @@
 import torch
 import pytest
-from .simple_model import SimpleModel
+from tests.unit.simple_model import SimpleModel
 from deepspeed import OnDevice
 from packaging import version as pkg_version
